+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## [2.9.11] - 2019-06-17
 ### Added
 - Compatibility with Base module's constructor changes
+### Fixed
+- Mapping of shipping method during order import
 
 ## [2.9.10] - 2019-06-04
 ### Fixed